Update Peer review
Great job finishing the challenge! 🎉

The looks awesome, for real!

A cherry on the top:

For a smoother color change of the links, you can use the transition property:
transition: background-color 0.5s, color 0.5s;
This ensures that the colors will change after 500 milliseconds, which makes it look ever better.

Keep up the grind. 💪